Move AutomationRule to projects app
