+++ readthedocs/builds/models.py
@@ -25,7 +25,6 @@ from readthedocs.builds.constants import BUILD_STATE_FINISHED
 from readthedocs.builds.constants import BUILD_STATE_TRIGGERED
 from readthedocs.builds.constants import BUILD_STATUS_CHOICES
 from readthedocs.builds.constants import BUILD_TYPES
-from readthedocs.builds.constants import CUSTOM_MATCH
 from readthedocs.builds.constants import EXTERNAL
 from readthedocs.builds.constants import EXTERNAL_VERSION_STATES
 from readthedocs.builds.constants import INTERNAL
@@ -1392,410 +1391,3 @@ class AutomationRuleMatch(TimeStampedModel):
 
     def get_action_past_tense(self):
         return self.ACTIONS_PAST_TENSE.get(self.action)
-
-
-class AutomationRule(TimeStampedModel):
-    """
-    Automation rule for versions (v2 design).
-
-    This is a redesigned automation rule system that combines version matching
-    with webhook filtering in a single, more flexible model.
-
-    The rule works as follows:
-    1. Checks if the version type matches one of the configured version_types
-    2. Checks if the version name matches the version_match_pattern (regex)
-    3. If webhook_filter is set, also checks the webhook data against webhook_match_pattern
-    4. If all checks pass, executes the configured action
-    """
-
-    ACTIVATE_VERSION_ACTION = "activate-version"
-    DELETE_VERSION_ACTION = "delete-version"
-    HIDE_VERSION_ACTION = "hide-version"
-    MAKE_VERSION_PUBLIC_ACTION = "make-version-public"
-    MAKE_VERSION_PRIVATE_ACTION = "make-version-private"
-    SET_DEFAULT_VERSION_ACTION = "set-default-version"
-    TRIGGER_BUILD_ACTION = "trigger-build"
-
-    ACTIONS = (
-        (ACTIVATE_VERSION_ACTION, _("Activate version")),
-        (HIDE_VERSION_ACTION, _("Hide version")),
-        (MAKE_VERSION_PUBLIC_ACTION, _("Make version public")),
-        (MAKE_VERSION_PRIVATE_ACTION, _("Make version private")),
-        (SET_DEFAULT_VERSION_ACTION, _("Set version as default")),
-        (DELETE_VERSION_ACTION, _("Delete version")),
-        (TRIGGER_BUILD_ACTION, _("Trigger build for version")),
-    )
-
-    VERSION_ACTIONS = (
-        ACTIVATE_VERSION_ACTION,
-        HIDE_VERSION_ACTION,
-        MAKE_VERSION_PUBLIC_ACTION,
-        MAKE_VERSION_PRIVATE_ACTION,
-        SET_DEFAULT_VERSION_ACTION,
-        DELETE_VERSION_ACTION,
-    )
-
-    BUILD_ACTIONS = (TRIGGER_BUILD_ACTION,)
-
-    WEBHOOK_FILTER_LABEL = "label"
-    WEBHOOK_FILTER_COMMIT_MESSAGE = "commit-message"
-    WEBHOOK_FILTER_FILE_PATTERN = "file-pattern"
-
-    WEBHOOK_FILTERS = (
-        WEBHOOK_FILTER_LABEL,
-        WEBHOOK_FILTER_COMMIT_MESSAGE,
-        WEBHOOK_FILTER_FILE_PATTERN,
-    )
-
-    TIMEOUT = 1  # timeout in seconds for regex matching
-
-    project = models.ForeignKey(
-        Project,
-        verbose_name=_("Project"),
-        related_name="automation_rules",
-        on_delete=models.CASCADE,
-    )
-
-    priority = models.PositiveIntegerField(
-        _("Rule priority"),
-        help_text=_("A lower number (0) means a higher priority"),
-        default=0,
-    )
-
-    description = models.CharField(
-        _("Description"),
-        max_length=255,
-        null=True,
-        blank=True,
-    )
-
-    # Version filtering
-    version_types = models.JSONField(
-        _("Version types"),
-        help_text=_(
-            "List of version types this rule applies to (e.g., ['tag', 'branch', 'external'])."
-        ),
-        default=list,
-    )
-
-    version_predefined_match_pattern = models.CharField(
-        _("Version predefined match pattern"),
-        help_text=_(
-            "Predefined pattern to match against the version name (e.g., 'semver-versions')"
-        ),
-        max_length=255,
-        choices=VERSION_PREDEFINED_MATCH_PATTERNS,
-    )
-
-    version_match_pattern = models.CharField(
-        _("Version match pattern"),
-        help_text=_("Regex pattern to match against the version name (e.g., '^release-.*')"),
-        max_length=255,
-        null=True,
-        blank=True,
-    )
-
-    # Webhook filtering (optional)
-    webhook_files_match_pattern = models.JSONField(
-        _("Webhook files match pattern"),
-        help_text=_(
-            "Pattern to match against the files changed in the webhook event. "
-            "Use fnmatch patterns (e.g., 'docs/*.md') and one pattern per line."
-        ),
-        max_length=1024,
-        null=True,
-        blank=True,
-    )
-    webhook_labels_match_pattern = models.CharField(
-        _("Webhook labels match pattern"),
-        help_text=_(
-            "Pattern to match against the labels in the webhook event. "
-            "Use a regex pattern to match against the labels (e.g., 'docs|build')."
-        ),
-        max_length=255,
-        null=True,
-        blank=True,
-    )
-    webhook_commit_message_match_pattern = models.CharField(
-        _("Webhook commit message match pattern"),
-        help_text=_(
-            "Pattern to match against the commit message in the webhook event. "
-            "Use a regex pattern to match against the commit message (e.g., 'fix|feature')."
-        ),
-        max_length=255,
-        null=True,
-        blank=True,
-    )
-
-    # Action to perform
-    action = models.CharField(
-        _("Action"),
-        help_text=_("Action to apply when the rule matches"),
-        max_length=32,
-        choices=ACTIONS,
-    )
-
-    enabled = models.BooleanField(
-        _("Enabled"),
-        default=True,
-        help_text=_("Whether this rule is enabled"),
-    )
-
-    _position_manager = ProjectItemPositionManager(position_field_name="priority")
-
-    class Meta:
-        ordering = ("priority", "-modified", "-created")
-        verbose_name = _("Automation rule")
-        verbose_name_plural = _("Automation rules")
-
-    def __str__(self):
-        return f"({self.priority}) {self.get_action_display()}"
-
-    def save(self, *args, **kwargs):
-        """Override method to update the other priorities before save."""
-        self._position_manager.change_position_before_save(self)
-        super().save(*args, **kwargs)
-
-    def delete(self, *args, **kwargs):
-        """Override method to update the other priorities after delete."""
-        super().delete(*args, **kwargs)
-        self._position_manager.change_position_after_delete(self)
-
-    def move(self, steps):
-        """
-        Change the priority of this Automation Rule.
-
-        This is done by moving it ``n`` steps,
-        relative to the other priority rules.
-        The priority from the other rules are updated too.
-
-        :param steps: Number of steps to be moved
-                      (it can be negative)
-        """
-        total = self.project.automation_rules.count()
-        current_priority = self.priority
-        new_priority = (current_priority + steps) % total
-        self.priority = new_priority
-        self.save()
-
-    def get_description(self):
-        if self.description:
-            return self.description
-        return f"{self.get_action_display()}"
-
-    def get_version_match_pattern(self):
-        # If a predefined match pattern is set, use it instead of the user-defined one.
-        # This allows us to have common patterns defined by us that users can easily select
-        # without needing to write regex themselves.
-        if self.version_predefined_match_pattern == CUSTOM_MATCH:
-            return self.version_match_pattern
-        version_predefined_match_pattern = VERSION_PREDEFINED_MATCH_PATTERN_VALUES.get(
-            self.version_predefined_match_pattern,
-        )
-        return version_predefined_match_pattern
-
-    def _match_version(self, version):
-        """
-        Check if the version matches this rule's version criteria.
-
-        :param version: Version instance to check
-        :return: True if the version matches, False otherwise
-        """
-        # Check version type
-        if version.type not in self.version_types:
-            return False
-
-        version_match_pattern = self.get_version_match_pattern()
-
-        # Check version name pattern
-        try:
-            match = regex.search(
-                version_match_pattern,
-                version.verbose_name,
-                flags=regex.VERSION0,
-                timeout=self.TIMEOUT,
-            )
-            return bool(match)
-        except TimeoutError:
-            log.warning(
-                "Timeout while parsing version name regex.",
-                pattern=version_match_pattern,
-                version_slug=version.slug,
-                rule_id=self.pk,
-            )
-            return False
-        except Exception:
-            log.exception(
-                "Error parsing version name regex.",
-                exc_info=True,
-            )
-            return False
-
-    def _match_webhook(self, changed_files=None, commit_message=None, labels=None):
-        """
-        Check if the webhook data matches this rule's webhook criteria.
-
-        All the webhook criteria that are set (files, commit message, labels)
-        must match for this to return True.
-
-        :param changed_files: List of file paths that were modified/added/deleted
-        :param commit_message: Commit message from the webhook event
-        :param labels: List of labels from PR webhook event
-        :return: True if all the webhook filter match, False otherwise
-        """
-
-        changed_files = changed_files or []
-        commit_message = commit_message or ""
-        labels = labels or []
-
-        # Use fnmatch matching for file paths
-        if self.webhook_files_match_pattern:
-            files_matched = False
-            for file_path in changed_files:
-                for file_pattern in self.webhook_files_match_pattern:
-                    if fnmatch.fnmatch(file_path, file_pattern):
-                        log.info(
-                            "File pattern matched for webhook rule.",
-                            file_path=file_path,
-                            pattern=file_pattern,
-                            rule_id=self.pk,
-                        )
-                        files_matched = True
-                        break
-            if not files_matched:
-                return False
-
-        # Use regex matching for commit message
-        if self.webhook_commit_message_match_pattern:
-            commit_matched = False
-            try:
-                for commit_pattern in self.webhook_commit_message_match_pattern:
-                    match = regex.search(
-                        commit_pattern,
-                        commit_message,
-                        flags=regex.VERSION0,
-                        timeout=self.TIMEOUT,
-                    )
-                    if match:
-                        log.info(
-                            "Commit message pattern matched for webhook rule.",
-                            commit_message=commit_message,
-                            pattern=commit_pattern,
-                            rule_id=self.pk,
-                        )
-                        commit_matched = True
-                        break
-            except TimeoutError:
-                log.warning(
-                    "Timeout while parsing regex for commit message.",
-                    pattern=commit_pattern,
-                    rule_id=self.pk,
-                )
-            except Exception:
-                log.exception(
-                    "Error parsing regex for commit message.",
-                    exc_info=True,
-                )
-            if not commit_matched:
-                return False
-
-        # Use regex matching for labels
-        if self.webhook_labels_match_pattern:
-            labels_matched = False
-            for label in labels:
-                for label_pattern in self.webhook_labels_match_pattern:
-                    try:
-                        match = regex.search(
-                            label_pattern,
-                            label,
-                            flags=regex.VERSION0,
-                            timeout=self.TIMEOUT,
-                        )
-                        if match:
-                            log.info(
-                                "Label pattern matched for webhook rule.",
-                                label=label,
-                                pattern=label_pattern,
-                                rule_id=self.pk,
-                            )
-                            labels_matched = True
-                            break
-                    except TimeoutError:
-                        log.warning(
-                            "Timeout while parsing regex for label.",
-                            pattern=label_pattern,
-                            rule_id=self.pk,
-                        )
-                    except Exception:
-                        log.exception(
-                            "Error parsing regex for label.",
-                            exc_info=True,
-                        )
-            if not labels_matched:
-                return False
-
-        return True
-
-    def match(self, version, changed_files=None, commit_message=None, labels=None):
-        """
-        Check if the version and webhook data match this rule's criteria.
-
-        :param version: Version instance to check
-        :param changed_files: List of file paths that were modified/added/deleted in the webhook event
-        :param commit_message: Commit message from the webhook event
-        :param labels: List of labels from PR webhook event
-        :return: True if the version and webhook data match the rule, False otherwise
-        """
-        if not self._match_version(version):
-            return False
-
-        if any(
-            [
-                changed_files,
-                commit_message,
-                labels,
-                self.webhook_files_match_pattern,
-                self.webhook_commit_message_match_pattern,
-                self.webhook_labels_match_pattern,
-            ]
-        ):
-            return self._match_webhook(changed_files, commit_message, labels)
-
-        return True
-
-    def run(self, version):
-        """
-        Run the rule.
-
-        :param version: Version instance to check and act upon
-        :return: True if the action was performed, False otherwise
-        """
-        actions_map = {
-            self.ACTIVATE_VERSION_ACTION: actions.activate_version,
-            self.HIDE_VERSION_ACTION: actions.hide_version,
-            self.MAKE_VERSION_PUBLIC_ACTION: actions.set_public_privacy_level,
-            self.MAKE_VERSION_PRIVATE_ACTION: actions.set_private_privacy_level,
-            self.SET_DEFAULT_VERSION_ACTION: actions.set_default_version,
-            self.DELETE_VERSION_ACTION: actions.delete_version,
-            self.TRIGGER_BUILD_ACTION: actions.trigger_build_for_version,
-        }
-
-        action_func = actions_map.get(self.action)
-        if action_func:
-            action_func(version)
-        else:
-            raise NotImplementedError(f"Action {self.action} is not implemented")
-
-        # TODO: we could consider adding more metadata here related to the webhook match.
-        # I'm not going to add that feature for now, but I'll consider to do it in a following iteration.
-        AutomationRuleMatch.objects.register_match(
-            rule=self,
-            version=version,
-        )
-        return True
-
-    def get_edit_url(self):
-        return reverse(
-            "projects_automation_rule_edit",
-            args=[self.project.slug, self.pk],
-        )
